Bind 32-bit ints correctly on big endian LP64 systems
The size of a long on 64-bit Linux, Unix, and other LP64 platforms
is 64-bit, so passing the address of it and saying it is a 32-bit
integer will result in only the top 4 bytes being read. While this
does work on little endian systems, on big endian systems these bytes
will always be 0 for values < INT_MAX.

Instead, change the type to be an int, which should be 32 bits on
all supported platforms.